requirements.txt
Flask==2.3.2
gunicorn==21.2.0
Pillow==10.0.0

requests==2.31.0